requirements.txt
discord.py==2.4.0
google-genai==0.2.2
Flask==2.2.3
